--- models/contact.js
@@ -1,24 +1,0 @@
-const { Schema, model } = require("mongoose");
-
-const contactSchema = Schema({
-     
-    name: {
-        type: String,
-        required: [true, 'Set name for contact'],
-    },
-    email: {
-        type: String,
-    },
-    phone: {
-        type: String,
-    },
-    favorite: {
-        type: Boolean,
-        default: false,
-    }
-});
-
-const Contact = model("contact", contactSchema);
-
-
-module.exports = Contact;
